Add notification system (Slack, Teams, Telegram, WhatsApp) + fix tests
- Add complete notification integrations for 4 platforms
- Add NOTIFICATIONS.md setup guide
- Fix test database isolation with SQLite StaticPool
- Fix Decimal JSON serialization in QuotaErrorResponse
- Fix middleware exception handling (UnboundLocalError)
- Fix test_add_user_to_team endpoint path
- All 61 tests passing

+++ app/models.py
@@ -330,6 +330,6 @@ class QuotaErrorResponse(SQLModel):
     """Error response for quota exceeded scenarios."""
     error: str
     code: str = "quota_exceeded"
-    quota_remaining: Decimal
+    quota_remaining: float  # Use float for JSON serialization
     message: str
     approval_process: dict
